refactor: repackage http request class
