+++ src/main/java/controller/Controller.java
@@ -1,6 +1,6 @@
 package controller;
 
-import http.HttpRequest;
+import http.request.HttpRequest;
 import http.response.HttpResponse;
 
 import java.io.IOException;
