+++ src/main/java/webserver/RequestHandler.java
@@ -1,6 +1,6 @@
 package webserver;
 
-import http.HttpRequest;
+import http.request.HttpRequest;
 import http.response.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
