+++ src/main/java/webserver/processor/Processor.java
@@ -1,6 +1,6 @@
 package webserver.processor;
 
-import http.HttpRequest;
+import http.request.HttpRequest;
 import http.response.HttpResponse;
 
 import java.io.IOException;
