+++ src/test/java/utils/LoginUtilTest.java
@@ -1,6 +1,6 @@
 package utils;
 
-import http.HttpRequest;
+import http.request.HttpRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
