+++ src/test/java/webserver/processor/ResourceProcessorTest.java
@@ -1,7 +1,7 @@
 package webserver.processor;
 
 
-import http.HttpRequest;
+import http.request.HttpRequest;
 import http.response.HttpResponse;
 import http.StatusCode;
 import org.junit.jupiter.api.BeforeEach;
