+++ src/test/java/webserver/processor/TemplateProcessorTest.java
@@ -1,6 +1,6 @@
 package webserver.processor;
 
-import http.HttpRequest;
+import http.request.HttpRequest;
 import http.response.HttpResponse;
 import http.StatusCode;
 import org.junit.jupiter.api.BeforeEach;
